multiboot2: simplify tag_type module
The tag_type module was holding way too many things. Now, there are smaller
modules. This should not be breaking as it is only an internal change.
